requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
docker==6.1.3
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
beautifulsoup4==4.12.2
psutil==5.9.6
requests>=2.28.0
langgraph>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
anthropic>=0.25.0
